refactor: LoadingSpinner 컴포넌트 commone 폴더로 이동 및 적용, 스토리북 적용
