+++ components/common/LoadingSpinner/LoadingSpinner.stories.tsx
@@ -1,0 +1,14 @@
+import type { Meta, StoryObj } from '@storybook/react';
+import LoadingSpinner from '.';
+
+const meta: Meta<typeof LoadingSpinner> = {
+  title: 'LoadingSpinner',
+  component: LoadingSpinner,
+};
+
+export default meta;
+type Story = StoryObj<typeof LoadingSpinner>;
+
+export const Default: Story = {
+  args: {},
+};
