+++ pages/search/projects.tsx
@@ -17,8 +17,8 @@ import useMutation from '@libs/client/useMutation';
 import { ProfileResponse } from '@libs/client/useUser';
 import client from '@libs/server/client';
 import Layout from '@components/common/Layout';
+import LoadingSpinner from '@components/common/LoadingSpinner';
 import HeadMeta from '@components/headMeta';
-import LoadingSpinner from '@components/loadingSpinner';
 import DeleteModal from '@components/profile/deleteModal';
 import ClickedProject from '@components/project/clickedProject';
 import ProjectItem from '@components/project/projectItem';
